pyproject.toml
[tool.poetry]
name = "JaskierGE"
version = "0.0.1"
description = "First Edition of Shiet"
authors = ["rishan <rishan1996@hotmail.co.uk>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.1.1"
youtube_dl = "^2021.12.17"
table2ascii = "^1.1.0"
python-dotenv = "^0.21.1"
PyNaCl = "^1.5.0"
asyncio = "^3.4.3"
ffmpeg-python = "^0.2.0"
yt-dlp = "^2023.7.6"
pydub = "^0.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
